FOSS #476: move Firebase out of :work so the foss APK links zero firebase-*
The foss runtime classpath still pulled firebase-auth + firebase-messaging
transitively via the shared :work module, even though foss runs keypair +
UnifiedPush at runtime. F-Droid rejects APKs that link proprietary deps, so
:work must be Firebase-free.

Relocate the Firebase-coupled classes from :work into the :app `google`
source set, leaving the Firebase-free seams/interfaces in :work:

- FirebaseRenewalAuthProvider -> app/src/google (com.rousecontext.app.auth),
  behind the existing RenewalAuthProvider seam (foss already binds
  KeypairRenewalAuthProvider).
- FcmReceiver (FirebaseMessagingService) -> app/src/google
  (com.rousecontext.app.push); its manifest entry already lived in
  app/src/google. It delegates to the Firebase-free WakeDispatcher /
  FcmTokenRegistrar, which stay in :work.
- TunnelForegroundService's per-connect FCM-token push (the last
  FirebaseMessaging usage) now goes through a new ConnectPushReporter seam in
  :work. google binds FcmConnectPushReporter (fetches the FCM token via the
  existing FcmTokenProvider seam and forwards it); foss binds a no-op (its
  UnifiedPush endpoint is reported by UnifiedPushBackgroundDelivery).

Drop firebase-auth/messaging from :work's build.gradle.kts; they remain
googleImplementation in :app. Moved tests to app/src/testGoogle.

google behavior is unchanged — same Firebase auth + FCM paths, just relocated.

Classpath evidence:
  fossDebugRuntimeClasspath  | grep firebase -> (nothing)
  googleDebugRuntimeClasspath | grep firebase -> firebase-auth, firebase-messaging, firebase-crashlytics

Part of #460. Fixes #476.
