+++ app/src/google/java/com/rousecontext/app/di/DistributionModule.kt
@@ -9,10 +9,12 @@ import com.rousecontext.app.auth.FirebaseAnonymousAuthClient
 import com.rousecontext.app.auth.FirebaseDeviceAuthTokenProvider
 import com.rousecontext.app.auth.FirebaseDeviceCredentialProvider
 import com.rousecontext.app.auth.FirebaseFcmTokenProvider
+import com.rousecontext.app.auth.FirebaseRenewalAuthProvider
 import com.rousecontext.app.delivery.BackgroundDelivery
 import com.rousecontext.app.delivery.NoOpBackgroundDelivery
+import com.rousecontext.app.push.FcmConnectPushReporter
 import com.rousecontext.app.support.FirebaseCrashReporter
-import com.rousecontext.work.FirebaseRenewalAuthProvider
+import com.rousecontext.work.ConnectPushReporter
 import com.rousecontext.work.RenewalAuthProvider
 import org.koin.dsl.module
 
@@ -37,11 +39,19 @@ val distributionModule = module {
     single<FcmTokenProvider> { FirebaseFcmTokenProvider() }
     single<DeviceAuthTokenProvider> { FirebaseDeviceAuthTokenProvider() }
 
+    // Per-connect push-target sync (issue #476): fetch the current FCM token and
+    // forward it to the relay each time the tunnel connects. The foss flavor
+    // binds a no-op since it reports a UnifiedPush endpoint instead.
+    single<ConnectPushReporter> {
+        FcmConnectPushReporter(tunnelClient = get(), tokenProvider = get())
+    }
+
     // Device-auth seam (issue #462): wraps the Firebase anon/ID-token clients.
     single<DeviceCredentialProvider> {
         FirebaseDeviceCredentialProvider(anonymousAuth = get(), deviceAuth = get())
     }
 
     // Cert-renewal auth provider (Firebase ID token + Keystore signature).
+    // Lives in the `google` source set (issue #476) so :work links no firebase-auth.
     single<RenewalAuthProvider> { FirebaseRenewalAuthProvider(signer = get()) }
 }
