+++ app/src/google/java/com/rousecontext/app/push/FcmReceiver.kt
@@ -1,9 +1,12 @@
-package com.rousecontext.work
+package com.rousecontext.app.push
 
 import android.util.Log
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
 import com.rousecontext.mcp.core.ProviderRegistry
+import com.rousecontext.work.FcmTokenRegistrar
+import com.rousecontext.work.TunnelForegroundService
+import com.rousecontext.work.WakeDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancel
@@ -11,17 +14,22 @@ import kotlinx.coroutines.launch
 import org.koin.android.ext.android.inject
 
 /**
- * Receives FCM messages (google flavor) and routes them through the shared
+ * Receives FCM messages (`google` flavor) and routes them through the shared
  * [WakeDispatcher]:
  *
  * - `type: "wake"` starts the [TunnelForegroundService] (only if integrations are enabled)
- * - `type: "renew"` enqueues a [CertRenewalWorker] via WorkManager (only if integrations are enabled)
+ * - `type: "renew"` enqueues a `CertRenewalWorker` via WorkManager (only if integrations are enabled)
  * - Unknown types are logged and ignored
  *
  * Wake and renewal requests are silently ignored when no integrations are enabled,
  * avoiding unnecessary foreground service starts and ACME cert quota usage. The
  * `foss` flavor's UnifiedPush receiver shares the same [WakeDispatcher] logic
  * (issue #463).
+ *
+ * `google`-flavor-only (issue #476): this `FirebaseMessagingService` lives in the `:app`
+ * `google` source set — alongside its flavor-scoped manifest entry — so the shared `:work`
+ * module links no `firebase-messaging`. The Firebase-free [WakeDispatcher] /
+ * [FcmTokenRegistrar] it delegates to stay in `:work`.
  */
 class FcmReceiver : FirebaseMessagingService() {
